test: Move test location to logical place.
